move nextPageQuery from LuceneDocCollectorTest to OrderedDocCollectorTest
LuceneDocCollectorTest is an integration test. To test the nextPageQuery
stuff the integration layer isn't required. So this commit moves them
into a unit test.
